pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hello-agents"
version = "0.1.1"
description = "灵活、可扩展的多智能体框架 - 基于Datawhale Hello-Agents教程"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "HelloAgents Team", email = "jjyaoao@126.com"}
]
keywords = ["agents", "llm", "ai", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/datawhalechina/hello-agents"
Repository = "https://github.com/datawhalechina/hello-agents"
Documentation = "https://github.com/datawhalechina/hello-agents"

[tool.setuptools.packages.find]
where = ["."]
include = ["hello_agents*", "agents*", "core*", "tools*", "utils*"]
exclude = ["tests*", "__pycache__*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
    "ignore::ResourceWarning",
]
